% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accdbfetch-package.R
\docType{package}
\name{accdbfetch-package}
\alias{accdbfetch}
\alias{accdbfetch-package}
\title{accdbfetch: A Package to Import tables from MS Access}
\description{
accdbfetch connects to an MS Access database using a file path supplied by user. After connection to the db is established, the function imports and tidys the imported tables and adds them to the R environment.
}
\author{
\strong{Maintainer}: Robert Cline \email{rccline@fastmail.fm} (\href{https://orcid.org/0000-0003-4758-3111}{ORCID})

}
\keyword{internal}
