% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accdb_fetch.R
\name{accdb_fetch}
\alias{accdb_fetch}
\title{Imports tables from MS Access using supplied path to Access DB.}
\usage{
accdb_fetch(x)
}
\arguments{
\item{x}{path to AccessDB}
}
\value{
tables from Access application.
}
\description{
Imports tables from MS Access using supplied path to Access DB.
}
\examples{
accdb_fetch("F:/D_Documents/_CheyenneCounty/NORSTAR/S Johnson Database/GTLS_SoJohnson Ver 1.0.accdb")
}
